Moved buffers into a buffers folder
