+++ CMakeLists.txt
@@ -18,7 +18,7 @@ include(C:/Users/unfea/Documents/vcpkg/vcpkg/scripts/buildsystems/vcpkg.cmake)
 include_directories(${CMAKE_SOURCE_DIR}/src/)
 
 add_library(Spark SHARED
-    "src/logging/Logger.cpp" "src/logging/Logger.h" "src/logging/Logging.cpp" "src/logging/Logging.h" "src/utils/DateUtils.h" "src/utils/DateUtils.cpp" "src/core/Application.cpp" "src/core/Application.h"  "src/Spark.h" "src/core/SparkEngine.h"   "src/logging/InternalLogging.h" "src/logging/InternalLogging.cpp" "src/events/Event.h" "src/events/EventHandler.h" "src/events/EventHandler.cpp" "src/events/EventListener.h" "src/events/EventListener.cpp"  "src/events/Events.h" "src/events/events/WindowResize.h" "src/events/EventContainer.h"  "src/pch.h"      "src/window/abstract/Window.h" "src/window/abstract/Window.cpp" "src/window/glfw/GLFWWindow.h" "src/window/glfw/GLFWWindow.cpp" "src/window/api/WindowFactory.h" "src/window/api/WindowFactory.cpp" "src/core/SparkEngine.cpp" "src/entities/BaseComponent.h" "src/entities/EntityManager.h" "src/entities/ComponentTypeRegistry.h" "src/entities/View.h" "src/entities/View.cpp" "src/systems/System.h" "src/systems/SystemManager.h" "src/systems/SystemManager.cpp" "src/graphics/abstract/Buffer.h"  "src/graphics/abstract/VertexBufferLayout.h"   "src/graphics/abstract/VertexBuffer.h" "src/graphics/abstract/IndexBuffer.h" "src/graphics/opengl/IndexBuffer.h" "src/graphics/opengl/IndexBuffer.cpp" "src/graphics/opengl/VertexBuffer.cpp" "src/graphics/opengl/VertexBuffer.h" "src/graphics/opengl/VertexArray.h" "src/graphics/opengl/VertexArray.cpp")
+    "src/logging/Logger.cpp" "src/logging/Logger.h" "src/logging/Logging.cpp" "src/logging/Logging.h" "src/utils/DateUtils.h" "src/utils/DateUtils.cpp" "src/core/Application.cpp" "src/core/Application.h"  "src/Spark.h" "src/core/SparkEngine.h"   "src/logging/InternalLogging.h" "src/logging/InternalLogging.cpp" "src/events/Event.h" "src/events/EventHandler.h" "src/events/EventHandler.cpp" "src/events/EventListener.h" "src/events/EventListener.cpp"  "src/events/Events.h" "src/events/events/WindowResize.h" "src/events/EventContainer.h"  "src/pch.h"      "src/window/abstract/Window.h" "src/window/abstract/Window.cpp" "src/window/glfw/GLFWWindow.h" "src/window/glfw/GLFWWindow.cpp" "src/window/api/WindowFactory.h" "src/window/api/WindowFactory.cpp" "src/core/SparkEngine.cpp" "src/entities/BaseComponent.h" "src/entities/EntityManager.h" "src/entities/ComponentTypeRegistry.h" "src/entities/View.h" "src/entities/View.cpp" "src/systems/System.h" "src/systems/SystemManager.h" "src/systems/SystemManager.cpp" "src/graphics/abstract/buffers/Buffer.h" "src/graphics/abstract/buffers/IndexBuffer.h" "src/graphics/abstract/buffers/VertexBuffer.h" "src/graphics/abstract/buffers/VertexBufferLayout.h" "src/graphics/abstract/buffers/VertexArray.h" "src/graphics/opengl/buffers/VertexBuffer.h" "src/graphics/opengl/buffers/IndexBuffer.h" "src/graphics/opengl/buffers/IndexBuffer.cpp" "src/graphics/opengl/buffers/VertexBuffer.cpp" "src/graphics/opengl/buffers/VertexArray.h" "src/graphics/opengl/buffers/VertexArray.cpp")
 
 # Spark Dependencies
 find_package(magic_enum CONFIG REQUIRED)
